# Conflicts:
#	src/features/analysis/ui/analysisAnalysis/analysisAnalysis.tsx
#	src/features/pakets/ui/userAnalysis.tsx

+++ src/features/pakets/ui/userAnalysis.tsx
@@ -7,8 +7,8 @@ import {IAnalysisProps, IUserPackets} from "../../../entities/pakets/model/paket
 interface IUserAnalysis {
     item: IAnalysisProps[],
     total: string,
-    onDeleteAnalysisId: (arg: number) => void,
-    onDeleteAllAnalysis: () => void
+    onDeleteAnalysisId?: (arg: number) => void,
+    onDeleteAllAnalysis?: () => void
 }
 
 export const UserAnalysis = memo(({item, onDeleteAnalysisId, onDeleteAllAnalysis, total}: IUserAnalysis) => {
@@ -29,14 +29,18 @@ export const UserAnalysis = memo(({item, onDeleteAnalysisId, onDeleteAllAnalysis
         //     analysisId: isActiveAnalysis,
         //     packagePrice: price
         // }))
-        onDeleteAnalysisId(isActiveAnalysis)
-        setIsDeleteAnalysis(false)
+        if (onDeleteAnalysisId){
+            onDeleteAnalysisId(isActiveAnalysis)
+            setIsDeleteAnalysis(false)
+        }
     }
 
     const onDeletePacket = () => {
         // dispatch(deletePacket(isActivePacket))
-        onDeleteAllAnalysis()
-        setIsDeletePacket(false)
+        if (onDeleteAllAnalysis){
+            onDeleteAllAnalysis()
+            setIsDeletePacket(false)
+        }
     }
 
     const onClickAnalysis = (id: number) => {
